+++ packages/astro/src/core/config.ts
@@ -13,6 +13,7 @@ import { BUNDLED_THEMES } from 'shiki';
 import { fileURLToPath, pathToFileURL } from 'url';
 import { mergeConfig as mergeViteConfig } from 'vite';
 import { z } from 'zod';
+import { LogOptions } from './logger/core.js';
 import { appendForwardSlash, prependForwardSlash, trimSlashes } from './path.js';
 import { arraify, isObject } from './util.js';
 
@@ -50,6 +51,9 @@ const ASTRO_CONFIG_DEFAULTS: AstroUserConfig & any = {
 		rehypePlugins: [],
 	},
 	vite: {},
+	legacy: {
+		astroFlavoredMarkdown: false,
+	},
 };
 
 async function resolvePostcssConfig(inlineOptions: any, root: URL): Promise<PostCSSConfigResult> {
@@ -89,7 +93,6 @@ export const LEGACY_ASTRO_CONFIG_KEYS = new Set([
 ]);
 
 export const AstroConfigSchema = z.object({
-	adapter: z.object({ name: z.string(), hooks: z.object({}).passthrough().default({}) }).optional(),
 	root: z
 		.string()
 		.optional()
@@ -124,6 +127,19 @@ export const AstroConfigSchema = z.object({
 		.union([z.literal('always'), z.literal('never'), z.literal('ignore')])
 		.optional()
 		.default(ASTRO_CONFIG_DEFAULTS.trailingSlash),
+	output: z
+		.union([z.literal('static'), z.literal('server')])
+		.optional()
+		.default('static'),
+	adapter: z.object({ name: z.string(), hooks: z.object({}).passthrough().default({}) }).optional(),
+	integrations: z.preprocess(
+		// preprocess
+		(val) => (Array.isArray(val) ? val.flat(Infinity).filter(Boolean) : val),
+		// validate
+		z
+			.array(z.object({ name: z.string(), hooks: z.object({}).passthrough().default({}) }))
+			.default(ASTRO_CONFIG_DEFAULTS.integrations)
+	),
 	build: z
 		.object({
 			format: z
@@ -150,14 +166,6 @@ export const AstroConfigSchema = z.object({
 			.optional()
 			.default({})
 	),
-	integrations: z.preprocess(
-		// preprocess
-		(val) => (Array.isArray(val) ? val.flat(Infinity).filter(Boolean) : val),
-		// validate
-		z
-			.array(z.object({ name: z.string(), hooks: z.object({}).passthrough().default({}) }))
-			.default(ASTRO_CONFIG_DEFAULTS.integrations)
-	),
 	style: z
 		.object({
 			postcss: z
@@ -172,9 +180,6 @@ export const AstroConfigSchema = z.object({
 		.default({}),
 	markdown: z
 		.object({
-			// NOTE: "mdx" allows us to parse/compile Astro components in markdown.
-			// TODO: This should probably be updated to something more like "md" | "astro"
-			mode: z.enum(['md', 'mdx']).default('mdx'),
 			drafts: z.boolean().default(false),
 			syntaxHighlight: z
 				.union([z.literal('shiki'), z.literal('prism'), z.literal(false)])
@@ -212,13 +217,23 @@ export const AstroConfigSchema = z.object({
 	vite: z
 		.custom<ViteUserConfig>((data) => data instanceof Object && !Array.isArray(data))
 		.default(ASTRO_CONFIG_DEFAULTS.vite),
+	legacy: z
+		.object({
+			astroFlavoredMarkdown: z
+				.boolean()
+				.optional()
+				.default(ASTRO_CONFIG_DEFAULTS.legacy.astroFlavoredMarkdown),
+		})
+		.optional()
+		.default({}),
 });
 
 /** Turn raw config values into normalized values */
 export async function validateConfig(
 	userConfig: any,
 	root: string,
-	cmd: string
+	cmd: string,
+	logging: LogOptions
 ): Promise<AstroConfig> {
 	const fileProtocolRoot = pathToFileURL(root + path.sep);
 	// Manual deprecation checks
@@ -325,7 +340,7 @@ export async function validateConfig(
 	const result = {
 		...(await AstroConfigRelativeSchema.parseAsync(userConfig)),
 		_ctx: {
-			pageExtensions: ['.astro', '.md'],
+			pageExtensions: ['.astro', '.md', '.html'],
 			scripts: [],
 			renderers: [],
 			injectedRoutes: [],
@@ -379,6 +394,7 @@ interface LoadConfigOptions {
 	flags?: Flags;
 	cmd: string;
 	validate?: boolean;
+	logging: LogOptions;
 }
 
 /**
@@ -450,7 +466,13 @@ export async function openConfig(configOptions: LoadConfigOptions): Promise<Open
 		userConfig = config.value;
 		userConfigPath = config.filePath;
 	}
-	const astroConfig = await resolveConfig(userConfig, root, flags, configOptions.cmd);
+	const astroConfig = await resolveConfig(
+		userConfig,
+		root,
+		flags,
+		configOptions.cmd,
+		configOptions.logging
+	);
 
 	return {
 		astroConfig,
@@ -496,7 +518,7 @@ export async function loadConfig(configOptions: LoadConfigOptions): Promise<Astr
 	if (config) {
 		userConfig = config.value;
 	}
-	return resolveConfig(userConfig, root, flags, configOptions.cmd);
+	return resolveConfig(userConfig, root, flags, configOptions.cmd, configOptions.logging);
 }
 
 /** Attempt to resolve an Astro configuration object. Normalize, validate, and return. */
@@ -504,10 +526,11 @@ export async function resolveConfig(
 	userConfig: AstroUserConfig,
 	root: string,
 	flags: CLIFlags = {},
-	cmd: string
+	cmd: string,
+	logging: LogOptions
 ): Promise<AstroConfig> {
 	const mergedConfig = mergeCLIFlags(userConfig, flags, cmd);
-	const validatedConfig = await validateConfig(mergedConfig, root, cmd);
+	const validatedConfig = await validateConfig(mergedConfig, root, cmd, logging);
 
 	return validatedConfig;
 }