+++ packages/astro/test/fixtures/static-build-ssr/astro.config.mjs
@@ -3,4 +3,5 @@ import nodejs from '@astrojs/node';
 
 export default defineConfig({
 	adapter: nodejs(),
+	output: 'server',
 });