+++ packages/astro/test/fixtures/tailwindcss/astro.config.mjs
@@ -1,9 +1,13 @@
 import { defineConfig } from 'astro/config';
 import tailwind from '@astrojs/tailwind';
+import mdx from '@astrojs/mdx';
 
 // https://astro.build/config
 export default defineConfig({
-	integrations: [tailwind()],
+	legacy: {
+		astroFlavoredMarkdown: true,
+	},
+	integrations: [tailwind(), mdx()],
 	vite: {
 		build: {
 			assetsInlineLimit: 0,