+++ packages/integrations/netlify/test/functions/base64-response.test.js
@@ -9,9 +9,7 @@ describe('Base64 Responses', () => {
 	before(async () => {
 		fixture = await loadFixture({
 			root: new URL('./fixtures/base64-response/', import.meta.url).toString(),
-			experimental: {
-				ssr: true,
-			},
+			output: 'server',
 			adapter: netlifyAdapter({
 				dist: new URL('./fixtures/base64-response/dist/', import.meta.url),
 				binaryMediaTypes: ['font/otf'],