+++ packages/integrations/vercel/src/static/adapter.ts
@@ -23,15 +23,7 @@ export default function vercelStatic(): AstroIntegration {
 				setAdapter(getAdapter());
 				_config = config;
 			},
-			'astro:build:start': async ({ buildConfig }) => {
-				if (String(process.env.ENABLE_VC_BUILD) !== '1') {
-					throw new Error(
-						`The enviroment variable "ENABLE_VC_BUILD" was not found. Make sure you have it set to "1" in your Vercel project.\nLearn how to set enviroment variables here: https://vercel.com/docs/concepts/projects/environment-variables`
-					);
-				}
-
-				buildConfig.staticMode = true;
-
+			'astro:build:start': async () => {
 				// Ensure to have `.vercel/output` empty.
 				// This is because, when building to static, outDir = .vercel/output/static/,
 				// so .vercel/output itself won't get cleaned.