+++ packages/integrations/svelte/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @astrojs/svelte
 
+## 0.5.0
+
+### Minor Changes
+
+- [#3570](https://github.com/withastro/astro/pull/3570) [`04070c0c1`](https://github.com/withastro/astro/commit/04070c0c12c00a3e17842ce48e36edc3f2c890a3) Thanks [@matthewp](https://github.com/matthewp)! - Bump to Vite 3!
+
+### Patch Changes
+
+- [#3993](https://github.com/withastro/astro/pull/3993) [`40a45e3ef`](https://github.com/withastro/astro/commit/40a45e3ef6284c024d442cf7cb8e36d8354a35d1) Thanks [@Princesseuh](https://github.com/Princesseuh)! - Fix optional props not being recognized properly in the editor
+
 ## 0.4.1
 
 ### Patch Changes